moved to new file, implemented in call_operation
